requirements.txt
absl-py==1.4.0
albumentations==1.3.0
alembic==1.6.5
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.3.4
astunparse==1.6.3
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
black==21.6b0
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cmake==3.26.1
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.2
distro==1.7.0
distro-info===1.1build1
docopt==0.6.2
docutils==0.19
duplicity==0.8.21
efficientnet-pytorch==0.7.1
entrypoints==0.4
fastapi==0.65.2
FastAPI-SQLAlchemy==0.2.1
fasteners==0.14.1
filelock==3.0.12
flatbuffers==23.3.3
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.4.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.51.3
h11==0.12.0
h5py==3.8.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.13.3
idna==3.3
imageio==2.27.0
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jax==0.4.6
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras==2.12.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
libclang==16.0.0
lit==16.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.4
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.0
notebook==6.4.8
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pathspec==0.8.1
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2021.5.29
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.5.0
ply==3.11
pretrainedmodels==0.7.4
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==4.22.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
pydantic==1.8.2
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.18.0
python-editor==1.0.4
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
qudida==0.0.4
readme-renderer==37.3
regex==2021.4.4
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
segmentation-models-pytorch==0.3.2
Send2Trash==1.8.1b0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.19
starlette==0.14.2
stevedore==5.0.0
sympy==1.11.1
systemd-python==234
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.13.1
testpath==0.5.0
testresources==2.0.1
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.12
toml==0.10.2
tomli==1.2.2
torch==2.0.0
torchaudio==2.0.1
torchinfo==1.7.2
torchvision==0.15.1
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
triton==2.0.0
twine==4.0.2
typing-extensions==3.10.0.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unittest2==1.1.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.14.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
